fix cache   NFE for big content length
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpclient/branches/4.6.x@1790038 13f79535-47bb-0310-9956-ffa450edef68

+++ httpclient-cache/src/main/java/org/apache/http/impl/client/cache/ResponseCachingPolicy.java
@@ -126,7 +126,7 @@ class ResponseCachingPolicy {
 
         final Header contentLength = response.getFirstHeader(HTTP.CONTENT_LEN);
         if (contentLength != null) {
-            final int contentLengthValue = Integer.parseInt(contentLength.getValue());
+            final long contentLengthValue = Long.parseLong(contentLength.getValue());
             if (contentLengthValue > this.maxObjectSizeBytes) {
                 return false;
             }
